background-color: update to use fullname and g prefix
Before this commit, background class selectors were prefixed `bgc`.

This is because `bc` stood for `border-color`. The three-character is
inconsist because `background` is one word.

Now, `gc` is the alias for `background-color`. I can't think of another
prefix this would collide with.

Additionally, colors now use full-name to match `color` class selectors.