update: opacity/luminocity rule to use '.' character
Before this commit, opacity and luminocity selectors looked like this:

    .o
    .o-5
    .o-0

This is because I really wanted to avoid characters that required
escapeing in selectors. But I realized that it's not big problem in
practice. These types of classes should be considered immutable and will
never need to be changed. So grepping for them is less of a
concern. Additionally, the chrome inspector shows the escaped rule.

Now, opacity and luminocity selectors look like this.
    o
    o.5
    o.0