update: b > bw
Before this commit, I was trying really hard to keep `border` a
single character to match it's box-model friends. Using it, it just
didn't make sense.

This is because the "one-character per dashed-delimited property" rule
is super strong. `b` felt totally out of place and broke the rule for no
good reason.

Now, `bw` is used to specify border-width.